Refactor the staticPageParser method
* Break down the method's logic
* Move it in on its semantically named methods
